requirements.txt
# ==== Core API Framework ====
fastapi==0.115.0
uvicorn[standard]==0.30.0

# ==== Machine Learning / NLP ====
transformers==4.44.2
torch==2.3.1
pandas==2.2.3
numpy==1.26.4

# ==== MongoDB ====
pymongo==4.7.3
certifi==2024.8.30
motor==3.4.0

# ==== Azure Cloud SDKs ====
azure-storage-blob==12.19.0
azure-identity==1.16.1

# ==== Utility / Optional ====
python-dotenv==1.0.1
requests==2.32.3
bcrypt==4.1.2
PyJWT==2.9.0
python-multipart==0.0.9
